test: move drift unit tests into src/commands/drift/__tests__
